+++ src/gluonts/torch/distributions/truncated_normal.py
@@ -11,26 +11,34 @@
 # express or implied. See the License for the specific language governing
 # permissions and limitations under the License.
 
-# mainly based from https://github.com/pytorch/rl/blob/main/torchrl/modules/distributions/continuous.py#L166
+# The implementation is strongly inspired from:
+# - https://github.com/pytorch/rl/blob/main/torchrl/modules/distributions/truncated_normal.py
+# - https://github.com/toshas/torch_truncnorm
 
+import math
 from numbers import Number
 from typing import Dict, Optional, Tuple, Union
 
 import torch
-from torch import distributions as D
 from torch.distributions import constraints
+from torch.distributions.utils import broadcast_all
 
 import torch.nn.functional as F
 from .distribution_output import DistributionOutput
 from gluonts.core.component import validated
 from torch.distributions import Distribution
-from .utils.truncated_normal import TruncatedNormal as _TruncatedNormal
 
-# speeds up distribution construction
-D.Distribution.set_default_validate_args(False)
+CONST_SQRT_2 = math.sqrt(2)
+CONST_INV_SQRT_2PI = 1 / math.sqrt(2 * math.pi)
+CONST_INV_SQRT_2 = 1 / math.sqrt(2)
+CONST_LOG_INV_SQRT_2PI = math.log(CONST_INV_SQRT_2PI)
+CONST_LOG_SQRT_2PI_E = 0.5 * math.log(2 * math.pi * math.e)
+torch.manual_seed(0)
 
 
-class TruncatedNormal(D.Independent):
+class TruncatedNormal(Distribution):
+    """Truncated Normal distribution."""
+
     """Implements a Truncated Normal distribution with location scaling.
 
     Location scaling prevents the location to be "too far" from 0, which ultimately
@@ -58,98 +66,166 @@ class TruncatedNormal(D.Independent):
         tanh_loc (bool, optional): if ``True``, the above formula is used for
             the location scaling, otherwise the raw value is kept.
             Default is ``False``;
+            
+    References: 
+    - https://people.sc.fsu.edu/~jburkardt/presentations/truncated_normal.pdf
+    
+    This implementation is strongly based on:
+        - https://github.com/pytorch/rl/blob/main/torchrl/modules/distributions/truncated_normal.py
+        - https://github.com/toshas/torch_truncnorm
     """
 
-    num_params: int = 2
-
     arg_constraints = {
-        "loc": constraints.real,
-        "scale": constraints.greater_than(1e-6),
+        "a": constraints.real,
+        "b": constraints.real,
     }
+    has_rsample = True
+    eps = 1e-6
 
     def __init__(
         self,
         loc: torch.Tensor,
         scale: torch.Tensor,
+        min: Union[torch.Tensor, float],
+        max: Union[torch.Tensor, float],
         upscale: Union[torch.Tensor, float] = 5.0,
-        min: Union[torch.Tensor, float] = -1.0,
-        max: Union[torch.Tensor, float] = 1.0,
         tanh_loc: bool = False,
     ):
-        err_msg = (
-            "TanhNormal max values must be strictly greater than min values"
-        )
-        if isinstance(max, torch.Tensor) or isinstance(min, torch.Tensor):
-            if not (max > min).all():
-                raise RuntimeError(err_msg)
-        elif isinstance(max, Number) and isinstance(min, Number):
-            if not max > min:
-                raise RuntimeError(err_msg)
-        else:
-            if not all(max > min):
-                raise RuntimeError(err_msg)
 
-        if isinstance(max, torch.Tensor):
-            self.non_trivial_max = (max != 1.0).any()
-        else:
-            self.non_trivial_max = max != 1.0
+        scale = scale.clamp_min(self.eps)
+        if tanh_loc:
+            loc = (loc / upscale).tanh() * upscale
+        loc = loc + (max - min) / 2 + min
 
-        if isinstance(min, torch.Tensor):
-            self.non_trivial_min = (min != -1.0).any()
-        else:
-            self.non_trivial_min = min != -1.0
-        self.tanh_loc = tanh_loc
-
-        self.device = loc.device
-        self.upscale = (
-            upscale
-            if not isinstance(upscale, torch.Tensor)
-            else upscale.to(self.device)
-        )
-
-        if isinstance(max, torch.Tensor):
-            max = max.to(self.device)
-        else:
-            max = torch.tensor(max, device=self.device)
-        if isinstance(min, torch.Tensor):
-            min = min.to(self.device)
-        else:
-            min = torch.tensor(min, device=self.device)
         self.min = min
         self.max = max
-        self.update(loc, scale)
-
-    def update(self, loc: torch.Tensor, scale: torch.Tensor) -> None:
-        if self.tanh_loc:
-            loc = (loc / self.upscale).tanh() * self.upscale
-        if self.non_trivial_max or self.non_trivial_min:
-            loc = loc + (self.max - self.min) / 2 + self.min
-        self.loc = loc
-        self.scale = scale
-
-        base_dist = _TruncatedNormal(
-            loc, scale, self.min.expand_as(loc), self.max.expand_as(scale)
+        self.loc, self.scale, a, b = broadcast_all(
+            loc, scale, self.min, self.max
         )
-        super().__init__(base_dist, 1, validate_args=False)
+        self._non_std_a = a
+        self._non_std_b = b
+        self.a = (a - self.loc) / self.scale
+        self.b = (b - self.loc) / self.scale
+
+        if isinstance(a, Number) and isinstance(b, Number):
+            batch_shape = torch.Size()
+        else:
+            batch_shape = self.a.size()
+
+        super(TruncatedNormal, self).__init__(batch_shape)
+        if self.a.dtype != self.b.dtype:
+            raise ValueError("Truncation bounds types are different")
+        if any(
+            (self.a >= self.b)
+            .view(
+                -1,
+            )
+            .tolist()
+        ):
+            raise ValueError("Incorrect truncation range")
+
+        eps = self.eps
+        self._dtype_min_gt_0 = eps
+        self._dtype_max_lt_1 = 1 - eps
+        self._little_phi_a = self._little_phi(self.a)
+        self._little_phi_b = self._little_phi(self.b)
+        self._big_phi_a = self._big_phi(self.a)
+        self._big_phi_b = self._big_phi(self.b)
+        self._Z = (self._big_phi_b - self._big_phi_a).clamp(eps, 1 - eps)
+        self._log_Z = self._Z.log()
+        little_phi_coeff_a = torch.nan_to_num(self.a, nan=math.nan)
+        little_phi_coeff_b = torch.nan_to_num(self.b, nan=math.nan)
+        self._lpbb_m_lpaa_d_Z = (
+            self._little_phi_b * little_phi_coeff_b
+            - self._little_phi_a * little_phi_coeff_a
+        ) / self._Z
+        self._mean = -(self._little_phi_b - self._little_phi_a) / self._Z
+        self._variance = (
+            1
+            - self._lpbb_m_lpaa_d_Z
+            - ((self._little_phi_b - self._little_phi_a) / self._Z) ** 2
+        )
+        self._entropy = (
+            CONST_LOG_SQRT_2PI_E + self._log_Z - 0.5 * self._lpbb_m_lpaa_d_Z
+        )
+        self._log_scale = self.scale.log()
+        self._mean_non_std = self._mean * self.scale + self.loc
+        self._variance_non_std = self._variance * self.scale**2
+        self._entropy_non_std = self._entropy + self._log_scale
+
+    @constraints.dependent_property
+    def support(self):
+        return constraints.interval(self.a, self.b)
 
     @property
-    def mode(self):
-        m = self.base_dist.loc
-        a = self.base_dist._non_std_a + self.base_dist._dtype_min_gt_0
-        b = self.base_dist._non_std_b - self.base_dist._dtype_min_gt_0
-        m = torch.min(torch.stack([m, b], -1), dim=-1)[0]
-        return torch.max(torch.stack([m, a], -1), dim=-1)[0]
+    def mean(self):
+        return self._mean_non_std
 
-    def log_prob(self, value, **kwargs):
-        a = self.base_dist._non_std_a + self.base_dist._dtype_min_gt_0
-        a = a.expand_as(value)
-        b = self.base_dist._non_std_b - self.base_dist._dtype_min_gt_0
-        b = b.expand_as(value)
-        value = torch.min(torch.stack([value, b], -1), dim=-1)[0]
-        value = torch.max(torch.stack([value, a], -1), dim=-1)[0]
-        return self.base_dist.log_prob(
-            value
-        )  # original: return super().log_prob(value, **kwargs)
+    @property
+    def variance(self):
+        return self._variance_non_std
+
+    @property
+    def entropy(self):
+        return self._entropy_non_std
+
+    @staticmethod
+    def _little_phi(x):
+        return (-(x**2) * 0.5).exp() * CONST_INV_SQRT_2PI
+
+    def _big_phi(self, x):
+        phi = 0.5 * (1 + (x * CONST_INV_SQRT_2).erf())
+        return phi.clamp(self.eps, 1 - self.eps)
+
+    @staticmethod
+    def _inv_big_phi(x):
+        return CONST_SQRT_2 * (2 * x - 1).erfinv()
+
+    def cdf_truncated_standard_normal(self, value):
+        return ((self._big_phi(value) - self._big_phi_a) / self._Z).clamp(0, 1)
+
+    def icdf_truncated_standard_normal(self, value):
+        y = self._big_phi_a + value * self._Z
+        y = y.clamp(self.eps, 1 - self.eps)
+        return self._inv_big_phi(y)
+
+    def log_prob_truncated_standard_normal(self, value):
+        return CONST_LOG_INV_SQRT_2PI - self._log_Z - (value**2) * 0.5
+
+    def _to_std_rv(self, value):
+        return (value - self.loc) / self.scale
+
+    def _from_std_rv(self, value):
+        return value * self.scale + self.loc
+
+    def cdf(self, value):
+        return self.cdf_truncated_standard_normal(self._to_std_rv(value))
+
+    def icdf(self, value):
+        sample = self._from_std_rv(self.icdf_truncated_standard_normal(value))
+
+        # clamp data but keep gradients
+        sample_clip = torch.stack(
+            [sample.detach(), self._non_std_a.detach().expand_as(sample)], 0
+        ).max(0)[0]
+        sample_clip = torch.stack(
+            [sample_clip, self._non_std_b.detach().expand_as(sample)], 0
+        ).min(0)[0]
+        sample.data.copy_(sample_clip)
+        return sample
+
+    def log_prob(self, value):
+        value = self._to_std_rv(value)
+        return self.log_prob_truncated_standard_normal(value) - self._log_scale
+
+    def rsample(self, sample_shape=None):
+        if sample_shape is None:
+            sample_shape = torch.Size([])
+        shape = self._extended_shape(sample_shape)
+        p = torch.empty(shape, device=self.a.device).uniform_(
+            self._dtype_min_gt_0, self._dtype_max_lt_1
+        )
+        return self.icdf(p)
 
 
 class TruncatedNormalOutput(DistributionOutput):
@@ -163,6 +239,9 @@ class TruncatedNormalOutput(DistributionOutput):
         upscale: float = 5.0,
         tanh_loc: bool = False,
     ) -> None:
+
+        assert min < max, "max must be strictly greater than min"
+
         super().__init__(self)
 
         self.min = min
@@ -174,9 +253,9 @@ class TruncatedNormalOutput(DistributionOutput):
             "scale": 1,
         }
 
-    # @classmethod
+    @classmethod
     def domain_map(
-        self,
+        cls,
         loc: torch.Tensor,
         scale: torch.Tensor,
     ):
